+++ doc/changelog.md
@@ -12,6 +12,7 @@
 
 ## Bugfixes
 
+- Fixed issue when agents got deleted which had completed at least one chunk
 - Error message which was always shown when adding new hash types fixed
 
 # v0.4.2 -> v0.4.3