+++ doc/changelog.md
@@ -13,6 +13,7 @@
 ## Bugfixes
 
 - Fixed issue when agents got deleted which had completed at least one chunk
+- Fixed conflicts on $_POST data agent vs. agentId
 - Error message which was always shown when adding new hash types fixed
 
 # v0.4.2 -> v0.4.3